+++ examples/document-search/multimodal.py
@@ -40,8 +40,8 @@ from ragbits.core.vector_stores.in_memory import InMemoryVectorStore
 from ragbits.document_search import DocumentSearch
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
 from ragbits.document_search.documents.sources import LocalFileSource
-from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
-from ragbits.document_search.ingestion.providers.dummy import DummyImageProvider
+from ragbits.document_search.ingestion.parsers.base import ImageDocumentParser
+from ragbits.document_search.ingestion.parsers.router import DocumentParserRouter
 
 IMAGES_PATH = Path(__file__).parent / "images"
 
@@ -68,7 +68,7 @@ async def main() -> None:
     vector_store_hybrid = HybridSearchVectorStore(vector_store_text, vector_store_image)
 
     # For this example, we want to skip OCR and make sure that we test direct image embeddings.
-    parser_router = DocumentProcessorRouter.from_config({DocumentType.JPG: DummyImageProvider()})
+    parser_router = DocumentParserRouter({DocumentType.JPG: ImageDocumentParser()})
 
     document_search = DocumentSearch(
         vector_store=vector_store_hybrid,