+++ packages/ragbits-core/tests/unit/prompts/test_parsers.py
@@ -8,7 +8,7 @@ from ragbits.core.prompt.parsers import ResponseParsingError
 from .test_prompt import _PromptOutput
 
 
-def test_prompt_with_str_output():
+async def test_prompt_with_str_output():
     """Test a prompt with a string output."""
 
     class TestPrompt(Prompt[None, str]):  # pylint: disable=unused-variable
@@ -17,10 +17,10 @@ def test_prompt_with_str_output():
         user_prompt = "Hello"
 
     prompt = TestPrompt()
-    assert prompt.parse_response("Hi") == "Hi"
+    assert await prompt.parse_response("Hi") == "Hi"
 
 
-def test_prompt_with_int_output():
+async def test_prompt_with_int_output():
     """Test a prompt with an int output."""
 
     class TestPrompt(Prompt[None, int]):  # pylint: disable=unused-variable
@@ -29,13 +29,13 @@ def test_prompt_with_int_output():
         user_prompt = "Hello"
 
     prompt = TestPrompt()
-    assert prompt.parse_response("1") == 1
+    assert await prompt.parse_response("1") == 1
 
     with pytest.raises(ResponseParsingError):
-        prompt.parse_response("a")
+        await prompt.parse_response("a")
 
 
-def test_prompt_with_model_output():
+async def test_prompt_with_model_output():
     """Test a prompt with a model output."""
 
     class TestPrompt(Prompt[None, _PromptOutput]):  # pylint: disable=unused-variable
@@ -44,15 +44,15 @@ def test_prompt_with_model_output():
         user_prompt = "Hello"
 
     prompt = TestPrompt()
-    assert prompt.parse_response('{"song_title": "Hello", "song_lyrics": "World"}') == _PromptOutput(
+    assert await prompt.parse_response('{"song_title": "Hello", "song_lyrics": "World"}') == _PromptOutput(
         song_title="Hello", song_lyrics="World"
     )
 
     with pytest.raises(ResponseParsingError):
-        prompt.parse_response('{"song_title": "Hello"}')
+        await prompt.parse_response('{"song_title": "Hello"}')
 
 
-def test_prompt_with_float_output():
+async def test_prompt_with_float_output():
     """Test a prompt with a float output."""
 
     class TestPrompt(Prompt[None, float]):  # pylint: disable=unused-variable
@@ -61,13 +61,13 @@ def test_prompt_with_float_output():
         user_prompt = "Hello"
 
     prompt = TestPrompt()
-    assert prompt.parse_response("1.0") == 1.0
+    assert await prompt.parse_response("1.0") == 1.0
 
     with pytest.raises(ResponseParsingError):
-        prompt.parse_response("a")
+        await prompt.parse_response("a")
 
 
-def test_prompt_with_bool_output():
+async def test_prompt_with_bool_output():
     """Test a prompt with a bool output."""
 
     class TestPrompt(Prompt[None, bool]):  # pylint: disable=unused-variable
@@ -76,14 +76,14 @@ def test_prompt_with_bool_output():
         user_prompt = "Hello"
 
     prompt = TestPrompt()
-    assert prompt.parse_response("true") is True
-    assert prompt.parse_response("false") is False
+    assert await prompt.parse_response("true") is True
+    assert await prompt.parse_response("false") is False
 
     with pytest.raises(ResponseParsingError):
-        prompt.parse_response("a")
+        await prompt.parse_response("a")
 
 
-def test_prompt_with_int_and_custom_parser():
+async def test_prompt_with_int_and_custom_parser():
     """Test a prompt with an int output and a custom parser."""
 
     class TestPrompt(Prompt[None, int]):  # pylint: disable=unused-variable
@@ -111,10 +111,10 @@ def test_prompt_with_int_and_custom_parser():
             raise ResponseParsingError("Could not parse response")
 
     prompt = TestPrompt()
-    assert prompt.parse_response("abcd k2") == 2
+    assert await prompt.parse_response("abcd k2") == 2
 
     with pytest.raises(ResponseParsingError):
-        prompt.parse_response("a")
+        await prompt.parse_response("a")
 
 
 def test_prompt_with_unknown_output_and_no_parser():
@@ -127,7 +127,7 @@ def test_prompt_with_unknown_output_and_no_parser():
             user_prompt = "Hello"
 
 
-def test_prompt_with_unknown_output_and_custom_parser():
+async def test_prompt_with_unknown_output_and_custom_parser():
     """Test a prompt with an output type that doesn't have a default parser but has a custom parser."""
 
     class TestPrompt(Prompt[None, list]):  # pylint: disable=unused-variable
@@ -152,5 +152,5 @@ def test_prompt_with_unknown_output_and_custom_parser():
             return response.split()
 
     prompt = TestPrompt()
-    assert prompt.parse_response("Hello World") == ["Hello", "World"]
-    assert prompt.parse_response("Hello") == ["Hello"]
+    assert await prompt.parse_response("Hello World") == ["Hello", "World"]
+    assert await prompt.parse_response("Hello") == ["Hello"]