+++ packages/ragbits-document-search/src/ragbits/document_search/ingestion/strategies/batched.py
@@ -1,16 +1,17 @@
 import asyncio
-from collections.abc import Iterable, Sequence
+from collections.abc import Iterable
 from dataclasses import dataclass
 from itertools import islice
 
 from ragbits.core.vector_stores.base import VectorStore
 from ragbits.document_search.documents.document import Document, DocumentMeta
-from ragbits.document_search.documents.element import Element, IntermediateElement
+from ragbits.document_search.documents.element import Element
 from ragbits.document_search.documents.sources import Source
-from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
-from ragbits.document_search.ingestion.intermediate_handlers.base import BaseIntermediateHandler
+from ragbits.document_search.ingestion.enrichers.router import ElementEnricherRouter
+from ragbits.document_search.ingestion.parsers.router import DocumentParserRouter
 from ragbits.document_search.ingestion.strategies.base import (
     IngestDocumentResult,
+    IngestError,
     IngestExecutionResult,
     IngestStrategy,
 )
@@ -19,11 +20,11 @@ from ragbits.document_search.ingestion.strategies.base import (
 @dataclass
 class IngestTaskResult:
     """
-    Represents the result of the document batch ingest tast.
+    Represents the result of the document batch ingest task.
     """
 
     document_uri: str
-    elements: Sequence[Element | IntermediateElement]
+    elements: list[Element]
 
 
 class BatchedIngestStrategy(IngestStrategy):
@@ -54,8 +55,8 @@ class BatchedIngestStrategy(IngestStrategy):
         self,
         documents: Iterable[DocumentMeta | Document | Source],
         vector_store: VectorStore,
-        parser_router: DocumentProcessorRouter,
-        enricher_router: dict[type[IntermediateElement], BaseIntermediateHandler],
+        parser_router: DocumentParserRouter,
+        enricher_router: ElementEnricherRouter,
     ) -> IngestExecutionResult:
         """
         Ingest documents sequentially in batches.
@@ -80,20 +81,20 @@ class BatchedIngestStrategy(IngestStrategy):
             successfully_parsed = [result for result in parse_results if isinstance(result, IngestTaskResult)]
             failed_parsed = [result for result in parse_results if isinstance(result, IngestDocumentResult)]
 
-            # Further split successful documents into intermediate and ready
-            intermediate_parsed = [
+            # Further split successful documents into to enrich and ready
+            to_enrich = [
                 result
                 for result in successfully_parsed
-                if any(isinstance(element, IntermediateElement) for element in result.elements)
+                if any(type(element) in enricher_router for element in result.elements)
             ]
             ready_parsed = [
                 result
                 for result in successfully_parsed
-                if not any(isinstance(element, IntermediateElement) for element in result.elements)
+                if not any(type(element) in enricher_router for element in result.elements)
             ]
 
-            # Enrich intermediate documents
-            enrich_results = await self._enrich_batch(intermediate_parsed, enricher_router)
+            # Enrich documents
+            enrich_results = await self._enrich_batch(to_enrich, enricher_router)
 
             # Split enriched documents into successful and failed
             successfully_enriched = [result for result in enrich_results if isinstance(result, IngestTaskResult)]
@@ -121,7 +122,7 @@ class BatchedIngestStrategy(IngestStrategy):
     async def _parse_batch(
         self,
         batch: list[DocumentMeta | Document | Source],
-        parser_router: DocumentProcessorRouter,
+        parser_router: DocumentParserRouter,
     ) -> list[IngestTaskResult | IngestDocumentResult]:
         """
         Parse batch of documents.
@@ -145,23 +146,34 @@ class BatchedIngestStrategy(IngestStrategy):
             ],
             return_exceptions=True,
         )
-        return [
-            IngestDocumentResult(
-                document_uri=uri,
-                error=response,
-            )
-            if isinstance(response, BaseException)
-            else IngestTaskResult(
-                document_uri=uri,
-                elements=response,
-            )
-            for uri, response in zip(uris, responses, strict=True)
-        ]
+
+        results: list[IngestTaskResult | IngestDocumentResult] = []
+        for uri, response in zip(uris, responses, strict=True):
+            if isinstance(response, BaseException):
+                if isinstance(response, Exception):
+                    results.append(
+                        IngestDocumentResult(
+                            document_uri=uri,
+                            error=IngestError.from_exception(response),
+                        )
+                    )
+                # Handle only standard exceptions, not BaseExceptions like SystemExit, KeyboardInterrupt, etc.
+                else:
+                    raise response
+            else:
+                results.append(
+                    IngestTaskResult(
+                        document_uri=uri,
+                        elements=response,
+                    )
+                )
+
+        return results
 
     async def _enrich_batch(
         self,
         batch: list[IngestTaskResult],
-        enricher_router: dict[type[IntermediateElement], BaseIntermediateHandler],
+        enricher_router: ElementEnricherRouter,
     ) -> list[IngestTaskResult | IngestDocumentResult]:
         """
         Enrich batch of documents.
@@ -177,25 +189,37 @@ class BatchedIngestStrategy(IngestStrategy):
             *[
                 self._call_with_error_handling(
                     self._enrich_elements,
-                    elements=[element for element in result.elements if isinstance(element, IntermediateElement)],
+                    elements=[element for element in result.elements if type(element) in enricher_router],
                     enricher_router=enricher_router,
                 )
                 for result in batch
             ],
             return_exceptions=True,
         )
-        return [
-            IngestDocumentResult(
-                document_uri=result.document_uri,
-                error=response,
-            )
-            if isinstance(response, BaseException)
-            else IngestTaskResult(
-                document_uri=result.document_uri,
-                elements=[element for element in result.elements if isinstance(element, Element)] + response,
-            )
-            for result, response in zip(batch, responses, strict=True)
-        ]
+
+        results: list[IngestTaskResult | IngestDocumentResult] = []
+        for result, response in zip(batch, responses, strict=True):
+            if isinstance(response, BaseException):
+                if isinstance(response, Exception):
+                    results.append(
+                        IngestDocumentResult(
+                            document_uri=result.document_uri,
+                            error=IngestError.from_exception(response),
+                        )
+                    )
+                # Handle only standard exceptions, not BaseExceptions like SystemExit, KeyboardInterrupt, etc.
+                else:
+                    raise response
+            else:
+                results.append(
+                    IngestTaskResult(
+                        document_uri=result.document_uri,
+                        elements=[element for element in result.elements if type(element) not in enricher_router]
+                        + response,
+                    )
+                )
+
+        return results
 
     async def _index_batch(
         self,
@@ -212,9 +236,7 @@ class BatchedIngestStrategy(IngestStrategy):
         Returns:
             The task results.
         """
-        elements = [
-            element for result in batch for element in result.elements if not isinstance(element, IntermediateElement)
-        ]
+        elements = [element for result in batch for element in result.elements]
         try:
             await self._call_with_error_handling(
                 self._remove_elements,
@@ -230,7 +252,7 @@ class BatchedIngestStrategy(IngestStrategy):
             return [
                 IngestDocumentResult(
                     document_uri=result.document_uri,
-                    error=exc,
+                    error=IngestError.from_exception(exc),
                 )
                 for result in batch
             ]